Moved preserve artifact method to NLPSaveStoreConnector
